backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.1.4
xarray==2023.11.1
rasterio==1.3.9
geopandas==0.14.1
shapely==2.0.2
folium==0.15.1
scikit-learn==1.3.2
torch==2.1.1
earthengine-api==0.1.366
opencv-python==4.8.1.78
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
redis==5.0.1
celery==5.3.4
python-multipart==0.0.6
python-dateutil==2.8.2
pytz==2023.3
netcdf4==1.6.5
h5py==3.10.0
reverse-geocoder==1.5.1

